requirements.txt
flask
python-dotenv 
jaclang 
byllm 
gitpython 
pydriller 
tree_sitter 
streamlit 
requests